Fix circular ref between BlobUtils and Handlers, move caching to FilerFileSystem
